added nav-paths constants
