--- src/app/core/constants/nav-excluded-paths.ts
@@ -1,1 +1,0 @@
-export const EXCLUDED_PATHS = ['/stock-checker'];
